refactor: set ListSkeleton together
