+++ src/components/Post/index.ts
@@ -1,3 +1,5 @@
 import Post from './Post';
+import PostList from './PostList';
 
 export default Post;
+export { PostList };
